[Core/Tiri] Harden headers with default member initialisers, [[nodiscard]], and debug assertions
ScopedObjectLock now uses default member initialisers for quicklock, error and obj, eliminating a GCC -Wmaybe-uninitialized false positive and removing the #ifdef __GNUC__ workaround.
Object struct gains [[nodiscard]] on query methods, overflow/underflow checks in pin()/unpin(), and freeIfReady() now returns bool.
Tiri thread code adds debug assertions for stack balance and refcount correctness.

+++ src/tiri/tiri_class.cpp
@@ -622,7 +622,7 @@ static ERR TIRI_Query(objScript *Self)
 
    if (not Self->ActivationCount) {
       // Announce once only to limit log noise
-      log.branch("Target: %d, Procedure: %s / ID #%" PRId64, Self->TargetID, Self->Procedure ? Self->Procedure : (STRING)".", (long long)Self->ProcedureID);
+      log.branch("Target: %d, Procedure: %s / ID #%" PRId64, Self->TargetID, Self->Procedure ? Self->Procedure : (STRING)".", Self->ProcedureID);
    }
 
    Self->CurrentLine = -1;
